requirements/price-list.txt
fpdf2
PyYAML
